Update Debian installation procedures in README.md
I took the liberty to update your excellent installation procedure to use `sudo apt install ./<package>` (note the **./**) which should automatically install all dependencies. In the end it should be possible to cleanly uninstall `cornrow` with `sudo apt purge cornrow` followed by `sudo apt autoremove --purge`.

I also tried to compile and build the package myself (because of #14) and found that your procedure was missing the packages `debhelper` and `libtool`. This was on a fairly standard Debian Buster installation. There was also a `cd cornrow` missing after untarring the source package.

Maybe that helps others to install the package …